TASK-83:build aggregation pipeline to calculate monthly sales and purchase number from user collection
