+++ backend/index.js
@@ -10,10 +10,7 @@ import marketPriceRoute from "./route/marketPriceRoute.js";
 import saleRoute from "./route/saleRoute.js";
 import purchaseRoute from "./route/purchaseRoute.js";
 import notificationRoute from "./route/notificationRoute.js";
-import salesFinRoute from "./route/salesFinRoute.js";
-import purchaseFinRoute from "./route/purchaseFinRoute.js";
-import userRoute_2 from "./route/userRoute_2.js";
-
+import tmpFinRoute from "./route/tmpFinRoute.js";
 
 const app = express();
 
@@ -26,7 +23,6 @@ app.listen(PORT, () => {
   console.log(`APP is listning to port :${PORT}`);
 });
 
-// app.use("/test", testRoute);
 app.use("/currentbalance", currentBalanceRoute);
 app.use("/user", userRoute);
 app.use("/inventory", inventoryRoute);
@@ -36,10 +32,9 @@ app.use("/marketprice", marketPriceRoute);
 app.use("/user", priceSuggestionRoute);
 app.use("/sale", saleRoute);
 app.use("/purchase", purchaseRoute);
-// app.use("/test", testRoute);
-app.use("/salesfin", salesFinRoute);
-app.use("/purchasefin", purchaseFinRoute);
-app.use("/user_2", userRoute_2);
+
+// This will be replaced by the endpoint 'user'
+app.use("/tmpFinRoute", tmpFinRoute);
 
 // connection to DB
 mongoose
