+++ backend/route/purchaseRoute.js
@@ -3,6 +3,7 @@ import * as purchaseController from "../controller/purchaseController.js";
 
 const router = express.Router({ mergeParams: true });
 router.get("/", purchaseController.getAllPurchase);
+router.get("/monthly-aggregate", purchaseController.aggregateMonthlyPurchases);
 router.get("/:id", purchaseController.getPurchaseById);
 router.post("/", purchaseController.createPurchase);
 router.patch("/:id", purchaseController.updatePurchase);
